Remove unused css-utils.ts
This file was created speculatively but isn't actually being used to
eliminate any duplication. The goal of this PR is to reduce existing
duplication, not add unused utility code.

If button class construction needs to be refactored in the future,
that should be a separate PR where the utility is actually used.